refactor: extract template card
