+++ apps/journeys-admin/src/components/TemplateCard/TemplateCard.tsx
@@ -15,9 +15,9 @@ import Skeleton from '@mui/material/Skeleton'
 import IconButton from '@mui/material/IconButton'
 import MoreVertIcon from '@mui/icons-material/MoreVert'
 import Link from 'next/link'
-import { GetPublishedTemplates_journeys as Journey } from '../../../../__generated__/GetPublishedTemplates'
-import { JourneyCardMenu } from '../../JourneyList/JourneyCard/JourneyCardMenu'
-import { StatusChip } from '../../JourneyList/JourneyCard/StatusChip'
+import { GetPublishedTemplates_journeys as Journey } from '../../../__generated__/GetPublishedTemplates'
+import { JourneyCardMenu } from '../JourneyList/JourneyCard/JourneyCardMenu'
+import { StatusChip } from '../JourneyList/JourneyCard/StatusChip'
 
 export interface TemplateCardProps {
   journey?: Journey
